fix: Remove issues created by b60170c on long dfs
Previously the property_mapping feature didn't work fully as intended
since long dataframes got split and that messed up the query dict
creation. Now long dataframes only get split after dict creation,
removing said issue.

Also adds a new module for general non-api related utilities.
